backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
reportlab==4.2.0
Pillow==10.3.0
pandas==2.2.2
python-dateutil==2.9.0
pydantic==2.7.1
